+++ srcs/Mobil.class.cpp
@@ -6,7 +6,7 @@
 //   By: niccheva <niccheva@student.42.fr>          +#+  +:+       +#+        //
 //                                                +#+#+#+#+#+   +#+           //
 //   Created: 2015/01/11 10:08:31 by niccheva          #+#    #+#             //
-//   Updated: 2015/01/11 10:14:13 by niccheva         ###   ########.fr       //
+//   Updated: 2015/01/11 14:03:00 by llapillo         ###   ########.fr       //
 //                                                                            //
 // ************************************************************************** //
 
@@ -33,7 +33,7 @@ Mobil const			&Mobil::operator=(Mobil const & rhs) {
 	return (*this);
 }
 
-void				Mobil::move(int x, int y) {
+void				Mobil::move(float x, float y) {
 	if (x >= 0)
 		this->_x = x;
 	if (y >= 0 )
